Changed Voice to Text
The main voice to text file is new, and is much more streamlined than the old one. I believe they both work, but the new one seems to run much faster, as it lacks all the clutter the old one has. Voice to Text alt has all of the old code.