Dual Changes
Made the scenarios into a dictionary so a scenario can be switched out with the sys variable at the top. Index.py is also updated to have a different host that hopefully allows for proper account functionality. To utilize the index.py changes, the second host link should be opened, not the first.